Summary - friday - 27th oct
Need to fix media queries and the display (flex). Missing prices on the movie list. The background color changes the footer - needs to be fixed. I also know I have repeated myself many times and some styles are not necessary. I have the layouts for every page - that is good.